+++ src/mod.ts
@@ -1,10 +1,10 @@
-import { yaclil } from "./yaclil.ts";
+import { arboj } from "./arboj.ts";
 import * as util from './util.ts';
 import * as std from './std.ts';
 import type * as types from '../types.d.ts';
 
 export {
-    yaclil,
+    arboj,
     util,
     std,
     type types,