+++ types.d.ts
@@ -1,5 +1,5 @@
 export type { Command, CommandDefinition, DynamicCommand } from './types/command';
 export type { Option } from './types/option';
-export type { YaclilOptions } from './types/init';
+export type { ArbojOptions } from './types/init';
 export type { ParsedCommand, ParsedOption, ParsedOptions, ParsedArgs, ParsedStandaloneOption, ParsedArg } from './types/parsed';
 export type { Arg } from './types/arg';